Add dl_video Tests
Added actual test code for the dl_video function. This test checks that the file is properly downloaded and named after a url is passed to the function. Also fixed the dl_video function since it did not pass these checks. Added a dictionary return that gives the title of the video and the file location as well as added identifiers to files to prevent issues when multiple videos with the same title are added.